Any time (minmax) player like I wanted & Ziv suggested. Has 1 serious issue - minmax worker continues on working after the timeout because the interrupt check occurs only once per minmax iteration (meaning the last iteration, which can take forever, will continue on long after the timeout, possibly entering the next move computation - which is very bad).

+++ ggp-base/src_my_player/minmax/MinMaxReporter.java
@@ -49,7 +49,7 @@ public class MinMaxReporter extends BasicReporter {
 
 	public void reportAndReset(Move selectedMove, int nodesInCache, int searchDepth, long duration) {
 		notifyObservers(new MinMaxEvent(selectedMove, exploredNodes, expandedNodes, prunedNodes, terminalNodes,
-				nodesInCache, cacheHits, searchDepth, sumBranchingFactor / expandedNodes, duration));
+				nodesInCache, cacheHits, searchDepth, sumBranchingFactor / (double) expandedNodes, duration));
 		resetCount();
 	}
